Build images based on the official Debian repository
All images are now based on the LLVM Debian repository.
This means that some image tags are downgraded to an older Debian version
(which is supported by the LLVM repo).

Additionally, all Docker images are generated and built in CI.
This should lead to faster and easier updates.

Note that builds for LLVM 3.3, 3.4 and 3.5 have been dropped.